+++ CHANGELOG.md
@@ -8,3 +8,7 @@
 ## 1.0.2
  - add function to save multiple files. (Useful in iOS because it will show dialog only once)
  - fix permission checking on Android 6 - 9.
+
+## 1.0.3
+ - check parameter and raise exception if invalid.
+ - automatically append number to filename if more than 2 files have the same name.